requirements.txt
Django==3.0.8
python-dotenv==0.14.0
psycopg2==2.8.5
djangorestframework==3.11.0
django-filter==2.3.0
drfapikey==0.0.3
django-extra-fields==2.0.5
Pillow==5.4.1
gunicorn==20.0.4